fixed some bugs when using the `lfg` related commands

+++ commands/stoplfg.js
@@ -9,12 +9,31 @@ const Discord = require("discord.js")
 const Mongoose = require('mongoose')
 
 module.exports = {
+    /**
+     * @param {Discord.Message} message 
+     * @param {String[]} args 
+     * @param {Discord.Client} client 
+     */
     run: async (message, args, client) => {
         /**
          * `Data` Model 
          * @type Mongoose.Model
          */
         const Data = require('../models/data')
+
+        /**
+         * If the member isn't looking for a group, return message
+         */
+        if(!message.member.roles.cache.find(role => role.name == lfg_role_name)) {
+            return message.channel.send( new Discord.MessageEmbed()
+                .setTitle(`Invalid action`)
+                .addField(
+                    `You are currently not looking for a group`,
+                    `You need to be looking for a group using \`${prefix}${require('./lfg').name}\` to be able to do this`,
+                    false
+                )
+            )
+        }
 
         /**
          * Find messages by the specified userID
@@ -24,19 +43,21 @@ module.exports = {
          * Delete the document from the database
          */
         try {
-            await Data.find({ userID: client.user.id }, (err, docs) => { 
+            await Data.find({ userID: message.author.id }, (err, docs) => { 
                 if(err) console.log(err)
-                docs.forEach(doc => {
-                    try {
-                        client.guilds.cache.find(guild => guild.id == doc.guildID).channels.cache.find(channel => channel.id == doc.channelID && channel.type == 'text').messages.cache.find(message => message.id == doc.messageID).delete() 
-                    }
-                    catch(err) {
-                        if(!err.stack.includes('Cannot read property \'delete\' of undefined'))
-                        console.log(err)
-                        console.log('\x1b[33mThe message that was looked for may not be cached anyomre\x1b[0m')
-                    }
-                    Data.findByIdAndDelete(doc._id, (err) => {if(err) console.log(err)})
-                })
+                if(docs) {
+                    docs.forEach(doc => {
+                        try {
+                            client.guilds.cache.find(guild => guild.id == doc.guildID).channels.cache.find(channel => channel.id == doc.channelID && channel.type == 'text').messages.cache.find(message => message.id == doc.messageID).delete() 
+                        }
+                        catch(err) {
+                            if(!err.stack.includes('Cannot read property \'delete\' of undefined'))
+                            console.log(err)
+                            console.log('\x1b[33mThe message that was looked for may not be cached anyomre\x1b[0m')
+                        }
+                        Data.findByIdAndDelete(doc._id, (err) => {if(err) console.log(err)})
+                    })
+                }
             }).catch(err => console.log(err))
         } catch (err) {
             console.log(err)
@@ -61,8 +82,8 @@ module.exports = {
             .then(msg => msg.delete({timeout: timeout}))
             .catch(err => console.log(err))
     },
-    name: "StopLookingForGroup",
-    alias: ["stoplfg"],
+    name: "stoplfg",
+    alias: ["stoplookingforgroup"],
     desc: "Marks the author as not LFG.",
     usage: `\`\`${prefix}stoplfg\`\``
 }
